requirements.txt
fastapi
uvicorn
azure-identity
azure-ai-projects
openai